+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Domain specific meta robots values support
 
 ### Changed
+ - Google Analytics ID domain specific cfg
  - domain specific meta robots and robots.txt in cfg
 
 ## [2.7.2] - 2017-12-09