+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.8.0] - 2017-12-17
+### Added
+ - Domain specific meta robots values support
+
+### Changed
+ - Domain specific Google Analytics ID in cfg
+ - Domain specific meta robots and robots.txt values in cfg
+
 ## [2.7.2] - 2017-12-09
 ### Fixed
  - Empty element <var> does not delete
@@ -280,6 +288,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Plugin LinkList creates and displays local variable ``linklist``.
 - System messages from previous request are now marked via class.
 
+[2.8.0]: https://bitbucket.org/igwr/cms/compare/v2.8.0..v2.7.2
 [2.7.2]: https://bitbucket.org/igwr/cms/compare/v2.7.2..v2.7.1
 [2.7.1]: https://bitbucket.org/igwr/cms/compare/v2.7.1..v2.7.0
 [2.7.0]: https://bitbucket.org/igwr/cms/compare/v2.7.0..v2.6.2